==> Step 1-3 Data Clean & Sentiment & Summarization.ipynb <==
### Step 1 Obatain and process data

import pandas as pd

file_path="https://huggingface.co/api/datasets/amazon_us_reviews/parquet/Music_v1_00/train/4.parquet"

df=pd.read_parquet(file_path)

# Size of original dataset 
len(df)

##### We will use 100,000 rows of data for this project

df=df[:100000]

df.columns #check which columns wont be needed

df.head(5)

df['marketplace'].unique()

# Keep necessary columns
columns_to_delete = ['marketplace', 'product_category', 'star_rating', 'helpful_votes', 'total_votes', 'vine', 'verified_purchase', 'review_date']

df.drop(columns=columns_to_delete, inplace=True)

df.head(5)

product_title=df['product_title'].unique() # How many distinct products
len(product_title)

len(df['product_id'].unique()) # Since the output will be recommended music title so we will use product_title instead of id going forward

### Step 2 Sentiment Analysis using textblob

!pip3 install textblob

from textblob import TextBlob

def get_sentiment(text):
    analysis = TextBlob(text)
    return analysis.sentiment.polarity

df['sentiment'] = df['review_body'].apply(get_sentiment) # Add sentiment column to store sentiment score

df['sentiment_label'] = df['sentiment'].apply(lambda x: 'positive' if x > 0.3 else 'negative' if x < 0.3 else 'neutral') # Define sentiment label of each row

df[df['sentiment_label']=='negative']

df.to_csv('sentiment_data.csv',index=False)

### Filter out overall positive feedback music

# Count the negative neutral positive reviews of each music
sentiment_counts=df.groupby('product_title')['sentiment_label'].value_counts().unstack(fill_value=0).sort_values(by='positive',ascending=False)

sentiment_counts

# Filter the music with only the most positive feedback to a new list
overall_positive = []
overall_negative=[]
for title in df['product_title'].unique():
    positive_count = sentiment_counts.loc[title, 'positive']
    negative_count = sentiment_counts.loc[title, 'negative']
    neutral_count = sentiment_counts.loc[title, 'neutral']
    if positive_count == max(positive_count,negative_count,neutral_count):
        overall_positive.append(title)
    else:
        overall_negative.append(title)

print(len(overall_positive))
print(len(overall_negative))

# Finalize the filtered dataframe with only music with positive feedbacks
filtered_df = df[(df['product_title'].isin(overall_positive)) & (df['sentiment_label'] == 'positive')]
len(filtered_df)


len(filtered_df['product_title'].unique()) # Size of distinct music remained

filtered_df['review_headline']

### Step 3 Positive reviews summarization with TextRank

# Combine all review body based on their product title
combined_reviews = filtered_df.groupby('product_title')['review_body'].apply(lambda x: '.'.join(x)).reset_index()

combined_reviews[70:90]

from pyspark.sql import SparkSession
from pyspark.sql.functions import udf
from pyspark.sql.types import StringType
from sumy.parsers.plaintext import PlaintextParser
from sumy.nlp.tokenizers import Tokenizer
import pyspark.pandas as ps
from sumy.summarizers.text_rank import TextRankSummarizer

# Initialize SparkSession
spark = SparkSession.builder \
    .appName("AppName") \
    .getOrCreate()
    
# Convert pandas DataFrame column to Spark DataFrame
spark_df = spark.createDataFrame(combined_reviews[['review_body']], schema=["sentence"])

# Use textrank to summarize the reviewbody

def summarize_textrank(reviews):
    NUM_SENTS = 1
    summary = str()
    parser = PlaintextParser.from_string(reviews, Tokenizer("english"))
    summarizer = TextRankSummarizer()
    sentences = summarizer(parser.document, NUM_SENTS)  # Summarize the document with 1 sentences
    for sentence in sentences:
        summary += (sentence.__unicode__())
    return summary

summarize_reviews_udf = udf(summarize_textrank, StringType())

# Add a new column with summarized reviews
spark_df=spark_df.withColumn("summary", summarize_reviews_udf(spark_df["sentence"]))

spark_df.show()

summary=spark_df.toPandas()

summary

combined_reviews['summary']=summary['summary']

print(len(combined_reviews['review_body'][7]))
print(len(combined_reviews['summary'][7]))


combined_reviews

combined_reviews.to_csv('data_w_summaries.csv',index=False)
